Rename instrumentations
