+++ src/instrumentations/aws-sdk-v2/attribute-extractors.test.ts
